src/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
google-cloud-datastore==2.4.0
google-cloud-storage==2.1.0
ndjson==0.3.1
Flask-Cors==1.10.3